Board now prints block coordinates
Wow, this took more work than it seemed like it should.

SudokuBoard now prints a DxD array of tuples.  Each tuple is the
coordinates (row, column) of that particular block.

Now that I know how to do this it should be a straight shot to
generating the blocks within each cell.