Mockup app: initialize, call out for initial board state
First, I'm adding a bunch of infrastructure:

- ArrayUtilities.js: convenience methods for dealing with 2D arrays
- SudokuUtilities.js: board-related utilities (get filled board,
      randomize board, delete entries)
- SudokuMockup.js: functions that impersonate server calls

Then the initial, incomplete drafts of game components:

- SudokuMain.js: application main container
- SudokuGame.js: game tree
- SudokuBoard.js: display for a single board

Right now, the app initializes, calls out for an initial board, and
then prints that board to the console.  Next step: pass the initial
board to SudokuGame.  That's an easy step.  After that I'm going to
have the board render assignments and move lists, which ought to take
a couple hours.

Once that rendering is going, I'm going to push the code and hand it to
Shelley and David so that we can start getting our pieces to talk to
one another.