Add reshape1Dto2D function for mogrifying JSON arrays
The JSON representation of a board is currently a 1D array.  This
might have been a design mistake, but right now it's the API, so I'll
run with it.

This function lets us take one of those 1D arrays and turn it back
into a 2D array that is more convenient to work with.