SudokuBoard can now display assignments
SudokuBoard now passes the assignment array and the move list array
down to all its functions.  When it gets to making squares, it will
print the assignment in a square if there is one.

Next up: print move lists if there are move lists.  In order to do this
I need to change the initialBoard mockup to actually return a non-full
board.