Board choice now propagates up to requestHeuristic call in SMain
As the top-level container, SudokuMain is responsible for marshaling
state and communicating with the game engine.  I've got callbacks
in place so that when the user selects a choice in an available cell
in the main board a call to requestHeuristic() occurs up in
SudokuMain.  That call doesn't do anything yet but will soon.