Serial number is now a property of the board object, not the component
I had been storing the serial number in the SudokuBoard component.  This is not the right approach.  SudokuBoard is just a renderer and an interaction front-end, not the board itself.

The serial number is now in the board itself.  It is assigned when the board is added to the game tree.