Refactor include/exclude, implement pivot, add listHeuristics
In order to implement the pivot heuristic, I needed for it to be simple
to say 'take this board and assign a value to a given cell'.  That's
the include/exclude heuristic.  I pulled its guts out and made it
a separate function.

That made it simple to write 'pivot': just a foreach loop and a few
checks.  I could make it simpler still if I used Array.prototype.map()
instead of a for loop.

I also put in a simple 'list heuristics' call that the client will
use to populate some of its interface.