first code from prev. sudoku library repo
Includes a sudoku library, some test puzzles, a few test json files,
working (at least on the simple test) pivot and assign actions,
automatic application of inclusion/exclusion, and all of the
implemented logical operators available.  Unfortunately there's
still quite a bit of tweaking to do to get it more complete for
all the interactions, but this should allow for pivoting on and
assigning a cell.