Starting to factor out game-tree rendering from SudokuGame
We need two separate renderers: one for the active board, one for
the game tree.  It is cleanest to have both of those separate from
SudokuGame, whose main concern is to keep track of the game tree and
the selected heuristic/logical operator.

This is work in progress.

If react-treebeard falls through, I'll need to cherry-pick this onto
another branch.