Use flask_cors to enable cross-origin requests
Our current design has the Javascript code being served from one
place and the game logic being served from someplace else.  This runs
into the definition of cross-origin requests.  As a result, we need
the flask-cors module (imported as flask_cors) to add the appropriate
header to grant permission.

This may change in the deployed version.  We'll see.