Decouple cell selection from action execution; highlight selection
The previous version of the code would immediately execute an operation
when the user clicked on a cell in the active board.  No longer.  Now
we just keep track of the selection and highlight it in the active board
display.

The execution will take place when the user clicks EXECUTE SELECTED
ACTION up in the interface.