Call out for boards given a game ID and pass the boards to SudokuGame
We're now hardcoding the game ID in SudokuMain.  In production, we'll
get that as one of our input parameters.

Right after SudokuMain renders for the first time, we call out to the
server to get a list of boards associated with this game ID.  We pass
that in for SudokuGame to use.