Checkpoint: toward up-front selection of logical operators
This is just a checkpoint.  Do not expect this commit to work.

I'm adding a new dialog box that will pop up when the user has to
select logical operators up front.  Then we need some logic in
SudokuGame (semi-documented for my own use in a couple comments) to
remember to pop up that dialog and make the selection.