Receive a list of puzzles from server instead of a list of puzzles names
We've simplified the startup sequence so that the server ships over
a list of puzzles (as JSON objects) instead of a list of puzzle names.
This also makes it much simpler to move from one puzzle to the next.
Instead of requiring multiple exchanges with the server (send over the
finished game tree, request the new board), we can just fire off the
finished game tree.

This will also wind up making it easier to reset the puzzle.