Add new component for button guarded by ok/cancel dialog
This component is right out of the React examples for the Dialog
component.  A button on the main page brings up a modal OK/Cancel
dialog.  If the user presses Cancel, the dialog disappears and no
action is taken.  If the user presses OK, the dialog disappears and
the button's original function is triggered.

I've modified the original component to expose most of the configurable
options as React props.