Rewrite SudokuMain as functional component to get access to useParams()
React Router allows me to get components of the URL with the useParams()
hook.  However, hooks can only be used in React functional components.
I wrote all of the components for this application as classes because
I didn't yet know about hooks (or the rest of the last 5 years of React
development).

I'm not going to go through the whole app and rewrite everything, not
this close to release, but I have to admit I'm tempted.