Add config files for nginx, launchd to run server and client on a Mac
nginx.conf: Default nginx config file with entries to route requests
to the Sudoku client and server.

sudoku-server.plist: launchd plist that runs sudoku server under
gunicorn and puts its socket in /tmp.

Opsec concerns are minimal.  The launchd plist reveals that Andy's
username is atwilso (which is public knowledge) and that his home
directory on one particular machine is at /Users/atwilso (which is
the default for a Mac, so not surprising).