Generate data files containing entirely legal characters
Our previous code generated data filenames containing spaces (legal but
inconvenient) and colons (may or may not be legal; check local laws and
filesystem documentation).

I've changed it to something without whitespace or colons.  I've also
added in code to sanitize the ID portion of the filename in case we ever
use this with user-supplied IDs instead of random integers.