Remove old cost-amortization code and replace with new.
For the pivot action: Cost of pivot gets amortized across
all N children of the action.

For assign, exclude:

Half the cost of assign/exclude gets charged to each child of the action.

Note that when you run logical operations in conjunction with any of
these actions, the operator runs *separately* (and incurs its cost) on
every single one of the children.  If you want to run the operator just on
the current board, use Execute Operations.

Next commit will include a message to that effect.