Add minWidth property to game board container
Most of the interface components are handled just fine by flexbox and MaterialUI's layout systems.  The game board is different: because of the shenanigans with borders and all those tiny little cells, I need to specify a minimum width to keep CSS from letting things overlap.

In this case, I had to drop down to raw CSS and add

style={{minWidth: "650px"}}

...to the grid item containing the board view.